Implement ssaerrors.DryRunImmutableError
This error type will be returned by jsondiff.Immutable when a PATCH
operation fails due to an immutable error from the api. The purpose of
the error type is to return to the caller a copy of the object that was
fetched from the cluster just prior to the immutable error being
recieved. An immutable error may be considered to be a valid-ish kind of
diff by the caller, and the caller may want a way to access the cluster
object that caused it. Because no diff is returned though, the normal
way of accessing the fetched object is not available.